Remove PDAL_HAVE dependencies from compression code.
