+++ pdal/compression/ZstdCompression.hpp
@@ -33,114 +33,41 @@
 ****************************************************************************/
 #pragma once
 
-#ifdef PDAL_HAVE_ZSTD
-
 #include "Compression.hpp"
-
-#include <zstd.h>
 
 namespace pdal
 {
 
+class ZstdCompressorImpl;
+
 class ZstdCompressor : public Compressor
 {
 public:
-    ZstdCompressor(BlockCb cb) : m_cb(cb)
-    {
-        m_strm = ZSTD_createCStream();
-        ZSTD_initCStream(m_strm, 15);
-    }
+    ZstdCompressor(BlockCb cb);
+    ~ZstdCompressor();
 
-    ~ZstdCompressor()
-        { ZSTD_freeCStream(m_strm); }
-
-
-    void compress(const char *buf, size_t bufsize)
-    {
-        m_inBuf.src = reinterpret_cast<const void *>(buf);
-        m_inBuf.size = bufsize;
-        m_inBuf.pos = 0;
-
-        size_t ret;
-        do
-        {
-            ZSTD_outBuffer outBuf
-                { reinterpret_cast<void *>(m_tmpbuf), CHUNKSIZE, 0 };
-            ret = ZSTD_compressStream(m_strm, &outBuf, &m_inBuf);
-            if (ZSTD_isError(ret))
-                break;
-            if (outBuf.pos)
-                m_cb(m_tmpbuf, outBuf.pos);
-        } while (m_inBuf.pos != m_inBuf.size);
-    }
-
-    void done()
-    {
-        size_t ret;
-        do
-        {
-            ZSTD_outBuffer outBuf
-                { reinterpret_cast<void *>(m_tmpbuf), CHUNKSIZE, 0 };
-            ret = ZSTD_endStream(m_strm, &outBuf);
-            if (ZSTD_isError(ret))
-                break;
-            if (outBuf.pos)
-                m_cb(m_tmpbuf, outBuf.pos);
-        } while (ret);
-    }
+    void compress(const char *buf, size_t bufsize);
+    void done();
 
 private:
-    BlockCb m_cb;
-
-    ZSTD_CStream *m_strm;
-    ZSTD_inBuffer m_inBuf;
-    char m_tmpbuf[CHUNKSIZE];
+    std::unique_ptr<ZstdCompressorImpl> m_impl;
 };
+
+class ZstdDecompressorImpl;
 
 class ZstdDecompressor : public Decompressor
 {
 public:
-    ZstdDecompressor(BlockCb cb) : m_cb(cb)
-    {
-        m_strm = ZSTD_createDStream();
-        ZSTD_initDStream(m_strm);
-    }
+    ZstdDecompressor(BlockCb cb);
+    ~ZstdDecompressor();
 
-    ~ZstdDecompressor()
-    {
-        ZSTD_freeDStream(m_strm);
-    }
-
-    void decompress(const char *buf, size_t bufsize)
-    {
-        m_inBuf.src = reinterpret_cast<const void *>(buf);
-        m_inBuf.size = bufsize;
-        m_inBuf.pos = 0;
-
-        size_t ret;
-        do
-        {
-            ZSTD_outBuffer outBuf
-                { reinterpret_cast<void *>(m_tmpbuf), CHUNKSIZE, 0 };
-            ret = ZSTD_decompressStream(m_strm, &outBuf, &m_inBuf);
-            if (ZSTD_isError(ret))
-                break;
-            if (outBuf.pos)
-                m_cb(m_tmpbuf, outBuf.pos);
-        } while (m_inBuf.pos != m_inBuf.size);
-    }
-
+    void decompress(const char *buf, size_t bufsize);
     void done()
     {}
 
 private:
-    BlockCb m_cb;
-
-    ZSTD_DStream *m_strm;
-    ZSTD_inBuffer m_inBuf;
-    char m_tmpbuf[CHUNKSIZE];
+    std::unique_ptr<ZstdDecompressorImpl> m_impl;
 };
 
 } // namespace pdal
 
-#endif // PDAL_HAVE_ZSTD
