+++ test/unit/CMakeLists.txt
@@ -58,7 +58,18 @@ PDAL_ADD_TEST(pdal_streaming_test FILES StreamingTest.cpp)
 PDAL_ADD_TEST(pdal_support_test FILES SupportTest.cpp)
 PDAL_ADD_TEST(pdal_utils_test FILES UtilsTest.cpp)
 PDAL_ADD_TEST(pdal_uuid_test FILES UuidTest.cpp)
-PDAL_ADD_TEST(pdal_compression_test FILES CompressionTest.cpp)
+if (PDAL_HAVE_LAZ_PERF)
+PDAL_ADD_TEST(pdal_lazperf_test FILES LazPerfTest.cpp)
+endif()
+if (PDAL_HAVE_ZLIB)
+PDAL_ADD_TEST(pdal_deflate_test FILES DeflateTest.cpp)
+endif()
+if (PDAL_HAVE_LZMA)
+PDAL_ADD_TEST(pdal_lzma_test FILES LzmaTest.cpp)
+endif()
+if (PDAL_HAVE_ZSTD)
+PDAL_ADD_TEST(pdal_zstd_test FILES ZstdTest.cpp)
+endif()
 
 #
 # sources for the native io
